Chore: Target Newer Android Builds

+++ app/src/main/java/com/google/android/diskusage/entity/FileSystemEntrySmall.java
@@ -37,7 +37,7 @@ public class FileSystemEntrySmall extends FileSystemEntry {
   }
 
   @Override
-  public FileSystemEntry filter(CharSequence pattern, int blockSize) {
+  public FileSystemEntry filter(CharSequence pattern, long blockSize) {
     return null;
   }
 }
