+++ app/src/main/java/com/google/android/diskusage/entity/FileSystemSystemSpace.java
@@ -20,12 +20,12 @@
 package com.google.android.diskusage.entity;
 
 public class FileSystemSystemSpace extends FileSystemSpecial {
-  public FileSystemSystemSpace(String name, long size, int blockSize) {
+  public FileSystemSystemSpace(String name, long size, long blockSize) {
     super(name, size, blockSize);
   }
 
   @Override
-  public FileSystemEntry filter(CharSequence pattern, int blockSize) {
+  public FileSystemEntry filter(CharSequence pattern, long blockSize) {
     return null;
   }
 }
